refactor: 컴포넌트 및 코드 스타일 통일
